Moved the Model to api folder
